Nene's Kitchen Tkinter App
The name of the GUI application is Nenes KItchen Ordering System. The purpose of the application is to provide a platform for customers to order food online. The goals of the application are to make ordering food easier and more efficient for customers, to provide customers with an easy-to-use user interface, and to help the company maximize profits by increasing customer satisfaction. The target audience is people of all ages who are interested in ordering food online. Also Bussiness who need catering.