Validation testing
validate_username: Checks if the entered username has a length between 3 and 20 characters.

validate_password: Checks if the entered password has a length between 8 and 50 characters.

validate_card_number: Checks if the entered card number is 13 to 16 digits long using a regular expression.

validate_cvv: Checks if the entered CVV is 3 or 4 digits long using a regular expression.